All of the users API calls are tested and working correctly
