+++ server/session/login.js
@@ -4,8 +4,8 @@ const users = require('../db/init').users
 const compare = require('./auth').compare
 
 /**
- * Uses a userId and a password to check if a user is signed in
- * @param userId - String -  The id of the user who we are trying to log in
+ * Uses a username and a password to check if a user is signed in
+ * @param username - String -  The username of the user who we are trying to log in
  * @param password - String - The password of the user who is trying to log in
  * @return {Promise}
  *  resolve - (Boolean || User Info Object) -
@@ -18,9 +18,9 @@ const compare = require('./auth').compare
  *      null if the user doesn't exist
  *  reject - (err) - Rejects if there is an error
  */
-function login (userId, password) {
+function login (username, password) {
   return new Promise((resolve, reject) => {
-    users.getUserById(userId).then((user) => {
+    users.getUserByName(username).then((user) => {
       if (!user || typeof user === 'undefined') {
         // If the team is undefined then there must not have been a match
         return resolve(null)
